Fix "comparison between signed/unsigned" warnings

+++ Classes/ShareKit/Core/Helpers/OAuth/Categories/NSMutableURLRequest+Parameters.m
@@ -72,7 +72,7 @@
 {
     NSMutableString *encodedParameterPairs = [NSMutableString stringWithCapacity:256];
     
-    int position = 1;
+    NSUInteger position = 1;
     for (OARequestParameter *requestParameter in parameters) 
 	{
         [encodedParameterPairs appendString:[requestParameter URLEncodedNameValuePair]];
