+++ Classes/ShareKit/Sharers/Services/Evernote/Helpers/thrift/transport/THTTPClient.m
@@ -101,7 +101,7 @@
 }
 
 
-- (int) readAll: (uint8_t *) buf offset: (int) off length: (int) len
+- (int) readAll: (uint8_t *) buf offset: (NSUInteger) off length: (NSUInteger) len
 {
   NSRange r;
   r.location = mResponseDataOffset;
